Updated Channel page. Change for Stream page.
Changed the channel banner placeholder image so that the Tx's are
smaller. Removed the followers section on the channel page. Video
section now takes up bottom half and has the same video loading features
as the video page.

Changed stream page so that when the App bar is opened it opens
instantly now.